test(data): Add DAO tests and test utilities
This commit introduces database tests for the TodoListUltimateDao to
validate core CRUD operations.

It also adds a utility file with dummy task data for testing and
disables schema export in the database configuration. A minor return
type fix in the DAO's insert method is also included.

+++ app/src/main/java/com/shekhargh/todolistultimate/data/TodoListUltimateDao.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.flow.Flow
 interface TodoListUltimateDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertItem(item: TodoTaskItem): Int
+    suspend fun insertItem(item: TodoTaskItem): Long
 
     @Delete
     suspend fun deleteItem(item: TodoTaskItem): Int
